Se refactorizo toda la estructura del panel de administrador a una version con mas escalabilidad y sostenimiento del mismo modulo, aplicando la estructura de atomic design para el propio modulo de admininstrador
