+++ app/(admin)/solicitudes.tsx
@@ -1,6 +1,13 @@
 import React from 'react';
-import { SolicitudesTable } from '../../components/admin';
+import { SolicitudesTable } from '../../modules/admin';
+import { ScrollView, View } from 'react-native';
 
 export default function SolicitudesManagementScreen() {
-  return <SolicitudesTable />;
+  return (
+    <ScrollView className="flex-1 bg-gray-50">
+      <View className="p-6">
+        <SolicitudesTable />
+      </View>
+    </ScrollView>
+  );
 }
