+++ modules/admin/Atoms/index.ts
@@ -1,0 +1,5 @@
+export { DocumentBadge } from './DocumentBadge';
+export { StatCard, type StatCard as StatCardType } from './StatCard';
+export { StatusBadge } from './StatusBadge';
+export { TypeBadge } from './TypeBadge';
+
