+++ modules/admin/Organisms/index.ts
@@ -1,0 +1,6 @@
+export { DocumentCard } from './DocumentCard';
+export { FilterPanel } from './FilterPanel';
+export { PropertyCard } from './PropertyCard';
+export { SolicitudCard } from './SolicitudCard';
+export { UserCard } from './UserCard';
+
